requirements.txt
pillow==11.0.0
PyPDF2==3.0.1
loguru==0.7.2
coverage==7.5.4
genbadge[coverage]==1.1.1
